Refactored permissions handling in saved photos screen
If user did not grant permission to read external
storage, button is now shown to request it.
Before, permission request was launched when user
navigated to saved photos screen, but it created
a bug, when user opened device settings screen and
was asked for storage permission, because of how
viewpager loads sibling screens.